Compiling version of the code

+++ src/preprocessing/preprocessE537.cc
@@ -208,8 +208,7 @@ namespace NangaParbat
             emit << YAML::Key << "values" << YAML::Value;
             emit << YAML::BeginSeq;
 
-
-            for (int j=0; j < bins.size(); j++)
+            for (int j = 0; j < (int) bins.size(); j++)
               {
                 if(bins[j] == 1)
                   {
